+++ src/util/timer.ts
@@ -1,8 +1,9 @@
-export interface TimerOptions {
+export interface TimerOptions<T = unknown> {
+    returnValue?: T | PromiseLike<T>;
     signal?: AbortSignal;
 }
 
-export const timeout = (ms: number, options?: TimerOptions): Promise<void> =>
+export const timeout = <T = void>(ms: number, options?: TimerOptions<T>): Promise<T> =>
     new Promise((resolve, reject) => {
         const signal = options?.signal;
         if(signal != null) {
@@ -15,5 +16,5 @@ export const timeout = (ms: number, options?: TimerOptions): Promise<void> =>
                 reject(signal.reason);
             });
         }
-        const id = setTimeout(resolve, ms);
+        const id = setTimeout(resolve, ms, options?.returnValue);
     });